rename the package, move compute stat to example
