--- packages/py_daymet/src/__init__.py
@@ -1,1 +1,0 @@
-from .main import download, get_dataset, compute_stat
