+++ packages/pydaymet/pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pydaymet"
+description = "Download data from Daymet portal"
+version = "0.0.1"
+readme = "README.md"
+license = { text = "Apache 2.0" }
+classifiers = [
+    "License :: OSI Approved :: Apache Software License",
+    "Programming Language :: Python",
+    "Intended Audience :: Science/Research",
+]
+keywords = [
+  "weather",
+]
+requires-python = ">=3.8"
+dependencies = [
+    "pyproj",
+]
