+++ packages/pydaymet/src/pydaymet/__init__.py
@@ -1,0 +1,1 @@
+from pydaymet.main import download, get_dataset, compute_stat
